# ci: Adopt Ruff for faster linting and formatting