--- a/.editorconfig
+++ b/.editorconfig
@@ -20,6 +20,10 @@ max_line_length = 0
 indent_style = space
 indent_size = 3
 
+[*.json]
+indent_style = space
+indent_size = 2
+
 [.travis.yml]
 indent_style = space
 indent_size = 2