--- a/changes/.misc.md
+++ b/changes/.misc.md
@@ -1,0 +1,1 @@
+Replaced Flake8 and isort with Ruff for faster linting and formatting