--- a/src/ai/backend/manager/api/cluster_template.py
+++ b/src/ai/backend/manager/api/cluster_template.py
@@ -15,9 +15,6 @@ from ai.backend.common.logging import BraceStyleAdapter
 from ..models import (
     TemplateType,
     UserRole,
-)
-from ..models import association_groups_users as agus
-from ..models import (
     domains,
     groups,
     keypairs,
@@ -25,6 +22,7 @@ from ..models import (
     session_templates,
     users,
 )
+from ..models import association_groups_users as agus
 from ..models.session_template import check_cluster_template
 from .auth import auth_required
 from .exceptions import InvalidAPIParameters, TaskTemplateNotFound