--- a/src/ai/backend/manager/api/groupconfig.py
+++ b/src/ai/backend/manager/api/groupconfig.py
@@ -11,9 +11,14 @@ from ai.backend.common import msgpack
 from ai.backend.common import validators as tx
 from ai.backend.common.logging import BraceStyleAdapter
 
-from ..models import MAXIMUM_DOTFILE_SIZE
+from ..models import (
+    MAXIMUM_DOTFILE_SIZE,
+    groups,
+    query_group_domain,
+    query_group_dotfiles,
+    verify_dotfile_name,
+)
 from ..models import association_groups_users as agus
-from ..models import groups, query_group_domain, query_group_dotfiles, verify_dotfile_name
 from .auth import admin_required, auth_required
 from .exceptions import (
     DotfileAlreadyExists,