--- a/src/ai/backend/manager/container_registry/base.py
+++ b/src/ai/backend/manager/container_registry/base.py
@@ -15,9 +15,8 @@ import trafaret as t
 import yarl
 
 from ai.backend.common.bgtask import ProgressReporter
-from ai.backend.common.docker import ImageRef, arch_name_aliases
+from ai.backend.common.docker import ImageRef, arch_name_aliases, validate_image_labels
 from ai.backend.common.docker import login as registry_login
-from ai.backend.common.docker import validate_image_labels
 from ai.backend.common.exception import InvalidImageName, InvalidImageTag
 from ai.backend.common.logging import BraceStyleAdapter
 from ai.backend.manager.models.image import ImageRow, ImageType