--- a/src/ai/backend/manager/models/alembic/versions/11146ba02235_change_char_col_to_str.py
+++ b/src/ai/backend/manager/models/alembic/versions/11146ba02235_change_char_col_to_str.py
@@ -8,7 +8,6 @@ Create Date: 2022-03-25 12:32:05.637628
 import sqlalchemy as sa
 from alembic import op
 from sqlalchemy.sql import text
-from sqlalchemy.sql.expression import bindparam
 
 # revision identifiers, used by Alembic.
 revision = "11146ba02235"