--- a/src/ai/backend/manager/models/alembic/versions/1f55a65cfc4f_add_status_column_to_vfolders.py
+++ b/src/ai/backend/manager/models/alembic/versions/1f55a65cfc4f_add_status_column_to_vfolders.py
@@ -9,7 +9,6 @@ import sqlalchemy as sa
 from alembic import op
 from sqlalchemy.dialects import postgresql
 
-from ai.backend.manager.models.base import EnumValueType
 from ai.backend.manager.models.vfolder import VFolderOperationStatus
 
 # revision identifiers, used by Alembic.