--- a/src/ai/backend/manager/models/alembic/versions/360af8f33d4e_merge_f83d_and_1f55.py
+++ b/src/ai/backend/manager/models/alembic/versions/360af8f33d4e_merge_f83d_and_1f55.py
@@ -5,8 +5,6 @@ Revises: f83d630c0bc9, 1f55a65cfc4f
 Create Date: 2022-10-02 12:03:01.116361
 
 """
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "360af8f33d4e"