--- a/src/ai/backend/manager/models/alembic/versions/3cf19d906e71_.py
+++ b/src/ai/backend/manager/models/alembic/versions/3cf19d906e71_.py
@@ -5,8 +5,6 @@ Revises: 22964745c12b, 5d8e6043455e
 Create Date: 2019-06-17 16:45:14.580560
 
 """
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "3cf19d906e71"