--- a/src/ai/backend/manager/models/alembic/versions/3f1dafab60b2_merge.py
+++ b/src/ai/backend/manager/models/alembic/versions/3f1dafab60b2_merge.py
@@ -5,8 +5,6 @@ Revises: c092dabf3ee5, 6f5fe19894b7
 Create Date: 2019-09-30 04:34:42.092031
 
 """
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "3f1dafab60b2"