--- a/src/ai/backend/manager/models/alembic/versions/95f51fc6ffdb_merge.py
+++ b/src/ai/backend/manager/models/alembic/versions/95f51fc6ffdb_merge.py
@@ -5,8 +5,6 @@ Revises: 6129745f49b6, 5bce905c21e5
 Create Date: 2022-11-24 08:40:16.719312
 
 """
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "95f51fc6ffdb"