--- a/src/ai/backend/manager/models/alembic/versions/9cd61b1ae70d_add_scheduable_field_to_agents.py
+++ b/src/ai/backend/manager/models/alembic/versions/9cd61b1ae70d_add_scheduable_field_to_agents.py
@@ -7,7 +7,6 @@ Create Date: 2020-07-01 15:02:13.979828
 """
 import sqlalchemy as sa
 from alembic import op
-from sqlalchemy.dialects import postgresql
 from sqlalchemy.sql.expression import true
 
 # revision identifiers, used by Alembic.