--- a/src/ai/backend/manager/models/alembic/versions/a7ca9f175d5f_merge.py
+++ b/src/ai/backend/manager/models/alembic/versions/a7ca9f175d5f_merge.py
@@ -5,8 +5,6 @@ Revises: d59ff89e7514, 11146ba02235
 Create Date: 2022-03-28 15:25:22.965843
 
 """
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "a7ca9f175d5f"