--- a/src/ai/backend/manager/models/alembic/versions/ce209920f654_create_task_template_table.py
+++ b/src/ai/backend/manager/models/alembic/versions/ce209920f654_create_task_template_table.py
@@ -7,10 +7,9 @@ Create Date: 2019-12-16 13:39:13.210996
 """
 import sqlalchemy as sa
 from alembic import op
-from sqlalchemy.dialects import postgresql as pgsql
 from sqlalchemy.sql import text
 
-from ai.backend.manager.models.base import GUID, ForeignKeyIDColumn, IDColumn
+from ai.backend.manager.models.base import GUID, IDColumn
 
 # revision identifiers, used by Alembic.
 revision = "ce209920f654"