--- a/src/ai/backend/manager/models/alembic/versions/d643752544de_.py
+++ b/src/ai/backend/manager/models/alembic/versions/d643752544de_.py
@@ -5,8 +5,6 @@ Revises: 51dddd79aa21, d2aafa234374
 Create Date: 2020-03-09 12:04:27.013567
 
 """
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "d643752544de"