--- a/src/ai/backend/manager/models/alembic/versions/da24ff520049_add_starts_at_field_into_kernels.py
+++ b/src/ai/backend/manager/models/alembic/versions/da24ff520049_add_starts_at_field_into_kernels.py
@@ -7,7 +7,6 @@ Create Date: 2020-06-18 20:47:22.152831
 """
 import sqlalchemy as sa
 from alembic import op
-from sqlalchemy.dialects import postgresql
 
 # revision identifiers, used by Alembic.
 revision = "da24ff520049"