--- a/src/ai/backend/manager/models/alembic/versions/e35332f8d23d_add_modified_at_to_users_and_kernels.py
+++ b/src/ai/backend/manager/models/alembic/versions/e35332f8d23d_add_modified_at_to_users_and_kernels.py
@@ -7,7 +7,6 @@ Create Date: 2020-07-01 14:02:11.022032
 """
 import sqlalchemy as sa
 from alembic import op
-from sqlalchemy.dialects import postgresql
 from sqlalchemy.sql.expression import bindparam
 
 from ai.backend.manager.models.base import IDColumn, convention