--- a/src/ai/backend/manager/models/alembic/versions/eec98e65902a_merge_with_vfolder_clone.py
+++ b/src/ai/backend/manager/models/alembic/versions/eec98e65902a_merge_with_vfolder_clone.py
@@ -5,8 +5,6 @@ Revises: d463fc5d6109, 97f6c80c8aa5
 Create Date: 2020-10-03 18:11:06.270486
 
 """
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "eec98e65902a"