This PR replaces flake8 and isort with [Ruff](https://beta.ruff.rs/docs/), which runs 100 times faster.
Fortunately, [Pantsbuild already has an integration for Ruff!](https://blog.pantsbuild.org/linting-python-at-warp-speed/)
(NOTE: You need to run `pants fix` instead of `pants fmt` to run Ruff's fixers.)
Still, black and mypy are required as separate tools.

This won't reduce the CI/CD time much because the most overheads come from the initialization of Pants, but will greatly improve the daily code writing experience with IDE and editors.

```
> time LOCKSET=ruff/3.11.4 ./py -m ruff check src
LOCKSET=ruff/3.11.4 ./py -m ruff check src  0.06s user 0.05s system 108% cpu 0.101 total
> time LOCKSET=flake8/3.11.4 ./py -m flake8 src
LOCKSET=flake8/3.11.4 ./py -m flake8 src  9.15s user 0.41s system 308% cpu 3.099 total
```

You may need to update your IDE and editor configurations, referring the updated developer guide in this PR.
(Once merged, it will be available at: https://docs.backend.ai/en/latest/dev/daily-workflows.html#using-ides-and-editors)

## Migration of an existing development setup

Run:
```shell
rm -rf dist/export/python/virtualenvs/flake8
rm -rf dist/export/python/virtualenvs/isort
pants export --resolve=ruff
```

And then reconfigure your IDE or editor settings.

<!-- readthedocs-preview sorna start -->
----
:books: Documentation preview :books:: https://sorna--1475.org.readthedocs.build/en/1475/

<!-- readthedocs-preview sorna end -->

<!-- readthedocs-preview sorna-ko start -->
----
:books: Documentation preview :books:: https://sorna-ko--1475.org.readthedocs.build/ko/1475/

<!-- readthedocs-preview sorna-ko end -->