@achimnol 2023-08-13
TODO: Need to merge the VSCode section in the dev guide with #1468.